Updating the priority of one requirement
This patch accomplishes two things, first it changes the priority
of one of the requirements from being necessary to be desirable.
We talked with dtroyer this morning and he thinks we don't need to
be an official project at first, as long as we are in the OpenStack
namespace and under the QA umbrella.

This patch also fixes the format of the file so lines of text are
not too big which makes the file harder to read (except on the links
which would break if modified).

Change-Id: Ia8b1a7623c2c81d5dca2e16b235dc45ad4163b66