Fix the numbering of the requirements
Looks like the last patch broke the formatting of the numbered
list in the requirements section. The last patch added line breaks
to the end of all lines in an attempt to make the line shorter for
developers or reviewers, but it looks like numbered lists cannot
have these line breaks or they are not recognized as lists. This
patch changes that back.

Change-Id: Ia35e28d2e6b879cf2498cd75bfe45838160953f0